requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.1.2
cfgv==3.3.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.3
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
filelock==3.3.1
identify==2.3.0
kombu==5.1.0
librabbitmq==2.0.0
Markdown==3.3.4
nodeenv==1.6.0
platformdirs==2.4.0
pre-commit==2.15.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
